Print max number of synapses at end of file
This is a workaround for post processing. I need to give Pandas an
estimate of the maximum 'width' of the dataframe.

The comma at the end is so that Pandas can load the line too, and then
I drop it manually. The skipfooter argument doesn't work with the 'c'
engine in read_csv, and the dtype argument doesn't work with the
'python' engine. I decided to go with the quicker c engine, so the
comma is a hack to accommodate it.