Store ceil of total numbers of synaptic elements
Storing floats causes the issue where the total number of elements can
be less than the number of connected elements (which are stored as
integers). But connectivity is only upadated if the difference between
the two is at least 1.

So, it makes sense to store the ceil which will either be equal or
greater.