Slow down dynamics of inhibitory post-synaptic elements
Most inhibitory synapses are formed on shafts, and in deprivation
studies, they did not observe a change in the dendritic architecture.
I'm taking this to mean that the formation and retraction of inhibitory
synapses is therefore dominated by the dynamics of the axonal bits.
Slowing down the dynamics of inhibitory dendritic elements implies that
inhibitory spines, which are a minority may retract, but largely, the
shafts remain present for synapse formation.