Pick targets at random for E* connections
They all weight the same, so one cannot make a decision based on weight.
Instead of it still calculating probabilities etc., save some
computation by picking random targets straight away.

I've used a separate function instead of using the random choices
directly because:

a) In the future, I may replace it with a different function
b) When deletion is weight based, the tuples are rows of [nid, w], so
the nid has to be fetched from these.