Correct synapse from post method
A neuron with free inhibitory elements can either be E or I and can
therefore connect with an I neuron to either form IE or II synapses
respectively. These two synapse classes are different and must be
created in different ways.

For a neuron with excitatory elements, this does not apply since it'll
create either EE or EI synapses which are identical parameter wise.