Define max probability for connection formation
This permits us to change the probability distribution for different
types of synapses. For example, E synapses are short range, while I
synapses are long range. The only way we can form long range connections
is by reducing the probability of connection formation for neurons near
to each other---since the probability function is NOT normalised (we've
taken this from Butz and van Ooyen), merely changing the width of the
Gaussian is not sufficient---the maximum probability must also be
reduced.